ENH: revisit encoding of algorithm in SR
Instead of encoding algorithm details for each individual feature in radiomics features
SR, do this at the level of the measurement group. This helps significantly reduce the
size of the SR content sequence, which should help work around the 1MB SQ limit of
Google Healthcare.

In the process of implementing this refinement, clean up the code to make it possible to
complete the test runs much faster, and make the code a bit more robust.